its/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
bcrypt==4.0.1
Django==4.2.7
django-cors-headers==4.3.0
django-db-connection-pool==1.2.4
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-standardized-errors==0.12.6
greenlet==3.0.1
mysqlclient==2.2.0
PyJWT==2.8.0
pytz==2023.3.post1
SQLAlchemy==2.0.23
sqlparams==5.1.0
sqlparse==0.4.4
typing-extensions==4.8.0
